refactor: move settings yaml facet helpers
