+++ packages/kernel/src/domain/settings-closeout.ts
@@ -1,4 +1,51 @@
 import { settingBlockValue } from "../layout/harness-settings.ts";
+import { setting } from "../layout/harness-settings.ts";
+
+function replaceScalar(body: string, indent: string, key: string, value: string): string {
+  const line = new RegExp(`^(${indent}${key}:[^\\S\\r\\n]*)[^#\\r\\n]*?([^\\S\\r\\n]*(?:#[^\\r\\n]*)?)$`, "mu");
+  if (!line.test(body)) throw new Error(`Missing ${key} in harness.yaml settings facet.`);
+  return body.replace(line, `$1${value}$2`);
+}
+
+export function replaceDefaultedScalar(
+  body: string,
+  indent: string,
+  key: string,
+  value: string,
+  fallback: string,
+): string {
+  if (setting(body, key) === undefined && value === fallback) return body;
+  return replaceScalar(body, indent, key, value);
+}
+
+export function replaceOptionalDefaultedScalar(
+  body: string,
+  indent: string,
+  key: string,
+  value: string,
+  fallback: string,
+): string {
+  if (setting(body, key) !== undefined) return replaceScalar(body, indent, key, value);
+  if (value === fallback) return body;
+  const header = /^settings:[^\r\n]*(?:\r?\n|$)/mu;
+  if (!header.test(body)) throw new Error("Missing settings block in harness.yaml.");
+  return body.replace(header, (match) => `${match}${indent}${key}: ${value}\n`);
+}
+
+export function replaceDefaultedBlockScalar(
+  body: string,
+  block: string,
+  key: string,
+  value: string,
+  fallback: string,
+): string {
+  if (settingBlockValue(body, block, key) === undefined && value === fallback) return body;
+  const section = new RegExp(`(^  ${block}:[^\\S\\r\\n]*(?:\\r?\\n))((?:    [^\\r\\n]*(?:\\r?\\n|$))*)`, "mu");
+  const match = section.exec(body);
+  if (!match) throw new Error(`Missing ${block} block in harness.yaml settings facet.`);
+  const replaced = replaceScalar(match[2]!, "    ", key, value);
+  return `${body.slice(0, match.index)}${match[1]}${replaced}${body.slice(match.index + match[0].length)}`;
+}
 
 export const closeoutProfiles = ["standard", "strict"] as const;
 export type CloseoutProfile = (typeof closeoutProfiles)[number];
